implemented setting card priority/color.
Changed dropdown menu to be attached to the clicked priority button element, instead of the HTML body element. Positioning logic remains unchanged, using .offset() of button element. The dropdown div is now simply added and removed from the DOM, instead of setting and unsetting its display property. This change was needed to let the dropdown menu have a way to reference the .card parent of the star button that was used to display it.
One further plan for the priority setting/color change is to do it through setting/unsetting classes on the .card element. This would let us specify the position of the card in the flexbox list by setting flex order. Top priority can then go to the top of the list, etc.